fix(search): stop the installed badge leaking across package sources
performSearch marked results installed by bare name against a name set
that mixes formulae, casks, and Mac App Store apps, so a cask hit named
like an installed formula (wireshark, for example) wrongly showed the
installed check.

Match on source-qualified IDs instead, which also stops a Mac App Store
app satisfying a same-named brew result. The search suite moves to its
own file to stay under the file-length ceiling.
